chore: adding enums for base currency and validations for getBuyUrl
