+++ packages/common/src/validations/index.ts
@@ -9,3 +9,4 @@ export * from "./sol";
 export * from "./eth";
 export * from "./friendship";
 export * from "./notification";
+export * from "./buy";
